+++ src/collections/infra/repositories/CollectionsRepository.ts
@@ -40,6 +40,7 @@ import { ReadError } from '../../../core/domain/repositories/ReadError'
 import { CollectionLinks } from '../../domain/models/CollectionLinks'
 import { CollectionSummary } from '../../domain/models/CollectionSummary'
 import { LinkingObjectType } from '../../domain/useCases/GetCollectionsForLinking'
+import { TemplateCreateDTO } from '../../domain/dtos/TemplateCreateDTO'
 
 export interface NewCollectionRequestPayload {
   alias: string
@@ -524,6 +525,20 @@ export class CollectionsRepository extends ApiRepository implements ICollections
           displayName: item.name
         }))
       })
+      .catch((error) => {
+        throw error
+      })
+  }
+
+  public async createTemplate(
+    collectionIdOrAlias: number | string,
+    template: TemplateCreateDTO
+  ): Promise<void> {
+    return this.doPost(
+      `/${this.collectionsResourceName}/${collectionIdOrAlias}/templates`,
+      template
+    )
+      .then(() => undefined)
       .catch((error) => {
         throw error
       })