+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog follows the principles of [Keep a Changelog](https://keepachangel
 ### Added
 
 - New Use Case: [Get Collections For Linking Use Case](./docs/useCases.md#get-collections-for-linking).
+- New Use Case: [Create a Template](./docs/useCases.md#create-a-template) under Collections.
 
 ### Changed